backend/requirements.txt
# Core FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangChain and AI/ML libraries
# Add PyTorch first to resolve dependency conflicts
torch>=2.0.0,<3.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.38
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tiktoken>=0.5.1

# Data validation and settings
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP client and file handling
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Vector databases
chromadb>=0.4.15

# Caching and task queue
redis>=5.0.1
celery>=5.3.4

# Database
SQLAlchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication and security
bcrypt>=4.1.2
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring and logging
loguru>=0.7.2
prometheus-client>=0.19.0

# BDD Testing
pytest-bdd>=7.0.1
pytest-asyncio>=0.23.2
pytest-cov>=4.0.0
factory-boy>=3.3.0
